packages/healthcare-core-mcp/pyproject.toml
[project]
name = "healthcare-core-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.63",
    "fastapi>=0.135.1",
    "fastmcp>=3.1.0",
    "mangum>=0.21.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]
